Revert "feat: ActiveTime 컴포넌트 하나로 통합. 훅을 prop으로 받아서 사용."
This reverts commit 18c16a1570696218b857398a8d4e2bd3b87ef821.
