+++ src/pages/@owner/estimate/Estimate.tsx
@@ -3,30 +3,31 @@ import { useNavigate } from 'react-router-dom';
 import Navigation from '@components/layout/navigation/Navigation';
 import { Icon } from '@components/icon/Icon';
 import Button from '@components/ui/button/Button';
-import { useEstimateForm, useEstimateTime } from '@pages/@owner/estimate/hooks';
+import { useEstimateForm } from '@pages/@owner/estimate/hooks';
 import {
   Food,
   Price,
   RegionSection,
   ActiveDate,
+  ActiveTime,
   NeedElectricity,
   Etc,
 } from '@pages/@owner/estimate/@section';
-import { FormProvider } from 'react-hook-form';
-import ActiveTime from '@components/active-time/ActiveTime';
 
 export default function Estimate() {
   const navigate = useNavigate();
   const {
-    methods,
     handleSubmit,
     formData,
     errors,
+    activeTime,
     isValid,
     updateLocation,
     updateDetailLocation,
     updateAvailableDateById,
     removeAvailableDateById,
+    updateStartActiveTime,
+    updateEndActiveTime,
     updateFood,
     updatePrice,
     updateNeedElectricity,
@@ -38,7 +39,7 @@ export default function Estimate() {
     navigate(-1);
   };
   return (
-    <FormProvider {...methods}>
+    <>
       <Navigation
         centerContent='예약 견적서 작성'
         leftIcon={<Icon name='ic_back' />}
@@ -59,7 +60,12 @@ export default function Estimate() {
           handleAddAvailableDate={handleAddAvailableDate}
           error={errors.availableDates}
         />
-        <ActiveTime useActiveTimeHook={useEstimateTime} />
+        <ActiveTime
+          activeTime={activeTime}
+          updateStartActiveTime={updateStartActiveTime}
+          updateEndActiveTime={updateEndActiveTime}
+          error={errors.activeTime}
+        />
         <Food
           food={formData.food}
           updateFood={updateFood}
@@ -87,6 +93,6 @@ export default function Estimate() {
           저장하기
         </Button>
       </footer>
-    </FormProvider>
+    </>
   );
 }
