+++ src/pages/@owner/estimate/hooks/use-time.ts
@@ -1,20 +1,25 @@
 import { useEffect, useState } from 'react';
-import { useFormContext } from 'react-hook-form';
-import type { FoodTruckFormData } from '@pages/@owner/food-truck-form/schemas/food-truck-form.schema';
-import { FOOD_TRUCK_ERROR_MESSAGE } from '@pages/@owner/food-truck-form/constants/food-truck';
+import type { UseFormSetError, UseFormSetValue } from 'react-hook-form';
+
+import type { EstimateFormData } from '@pages/@owner/estimate/utils/estimate.schema';
 import type { TimeType } from '@type/time-types';
+import { ESTIMATE_ERROR_MESSAGE } from '@pages/@owner/estimate/constants/estimate';
 
-export const useFoodTruckFormTime = () => {
-  const {
-    setValue,
-    watch,
-    formState: { errors },
-    setError,
-  } = useFormContext<FoodTruckFormData>();
+//time, date 관련 로직
+interface UseEstimateTimeProps {
+  formData: EstimateFormData;
+  setValue: UseFormSetValue<EstimateFormData>;
+  setError: UseFormSetError<EstimateFormData>;
+}
 
-  const formActiveTime = watch('activeTime');
-  const formTimeDiscussRequired = watch('timeDiscussRequired');
+export const useEstimateTime = ({
+  formData,
+  setValue,
+  setError,
+}: UseEstimateTimeProps) => {
+  const formActiveTime = formData.activeTime;
 
+  //time 관련 로직
   const [startActiveTime, setStartActiveTime] = useState<TimeType | null>(null);
   const [endActiveTime, setEndActiveTime] = useState<TimeType | null>(null);
 
@@ -38,20 +43,20 @@ export const useFoodTruckFormTime = () => {
     }
     if (!startActiveTime) {
       setError('activeTime', {
-        message: FOOD_TRUCK_ERROR_MESSAGE.activeTime.start,
+        message: ESTIMATE_ERROR_MESSAGE.activeTime.start,
       });
       return;
     }
     if (!endActiveTime) {
       setError('activeTime', {
-        message: FOOD_TRUCK_ERROR_MESSAGE.activeTime.end,
+        message: ESTIMATE_ERROR_MESSAGE.activeTime.end,
       });
       return;
     }
     if (startActiveTime && endActiveTime) {
       if (isInvalidTimeRange(startActiveTime, endActiveTime)) {
         setError('activeTime', {
-          message: FOOD_TRUCK_ERROR_MESSAGE.activeTime.invalid,
+          message: ESTIMATE_ERROR_MESSAGE.activeTime.invalid,
         });
         return;
       }
@@ -104,24 +109,10 @@ export const useFoodTruckFormTime = () => {
     });
   }, [formActiveTime]);
 
-  const updateTimeDiscussRequired = (timeDiscussRequired: boolean) => {
-    setValue('timeDiscussRequired', timeDiscussRequired, {
-      shouldValidate: true,
-    });
-  };
-
   return {
-    // Data
     startActiveTime,
     endActiveTime,
-    timeDiscussRequired: formTimeDiscussRequired,
-
-    // Errors
-    activeTimeError: errors.activeTime?.message,
-
-    // Actions
     updateStartActiveTime,
     updateEndActiveTime,
-    updateTimeDiscussRequired,
   };
 };
