+++ src/pages/@owner/food-truck-form/@section/time-section/index.tsx
@@ -1,1 +1,2 @@
+export { default as ActiveTime } from '@pages/@owner/food-truck-form/@section/time-section/ActiveTime';
 export { default as ActiveDate } from '@pages/@owner/food-truck-form/@section/time-section/ActiveDate';
