+++ src/pages/@owner/food-truck-form/FoodTruckForm.tsx
@@ -12,7 +12,10 @@ import {
   FoodTruckOption,
   FoodTruckPhoto,
 } from '@pages/@owner/food-truck-form/@section/basic-info-section/index';
-import { ActiveDate } from '@pages/@owner/food-truck-form/@section/time-section/index';
+import {
+  ActiveTime,
+  ActiveDate,
+} from '@pages/@owner/food-truck-form/@section/time-section/index';
 import {
   AvailableQuantity,
   NeedElectricity,
@@ -24,8 +27,6 @@ import MenuInfo from '@pages/@owner/food-truck-form/@section/menu-section/MenuIn
 import { useFoodTruckForm } from '@pages/@owner/food-truck-form/hooks/use-food-truck-form';
 import useFoodTruckDetail from '@pages/food-truck-detail/hooks/use-food-truck-detail';
 import { ROUTES } from '@router/constant/routes';
-import ActiveTime from '@components/active-time/ActiveTime';
-import { useFoodTruckFormTime } from '@pages/@owner/food-truck-form/hooks/use-food-truck-form-time';
 
 // 메인 컴포넌트
 export default function FoodTruckForm() {
@@ -64,7 +65,7 @@ export default function FoodTruckForm() {
         <FoodTruckName />
         <FoodTruckDescription />
         <FoodTruckPhoneNumber />
-        <ActiveTime useActiveTimeHook={useFoodTruckFormTime} />
+        <ActiveTime />
         <RegionSection foodTruckId={foodTruckId} />
 
         <MenuCategory />
